// typescript/types.ts
/* ----------------------------------------------------------------------------------------------------------------------------------------------------- */
/* Select Option Type */
/* ----------------------------------------------------------------------------------------------------------------------------------------------------- */
type SelectOption = {
    key: KeyType;
    value: String|Number;
    text: String;
};
/* ----------------------------------------------------------------------------------------------------------------------------------------------------- */
/* Exports */
/* ----------------------------------------------------------------------------------------------------------------------------------------------------- */
export type { SelectOption };